FBR-524 deleting using instanceOf, create a abstract method to setupActionBar on screens that it’s needed
