+++ FBReader/src/main/java/org/geometerplus/android/fbreader/preferences/ZLPreferenceActivity.java
@@ -19,38 +19,25 @@
 
 package org.geometerplus.android.fbreader.preferences;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 
 import android.app.Dialog;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.*;
 import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
-import android.view.WindowManager;
 import android.widget.LinearLayout;
 
 import org.benetech.android.R;
-import org.geometerplus.android.fbreader.network.bookshare.Bookshare_Webservice_Login;
-import org.geometerplus.fbreader.fbreader.ActionCode;
-import org.geometerplus.fbreader.fbreader.SyncReadingListsWithBookshareAction;
-import org.geometerplus.zlibrary.core.application.ZLApplication;
-import org.geometerplus.zlibrary.core.library.ZLibrary;
 import org.geometerplus.zlibrary.core.options.*;
 import org.geometerplus.zlibrary.core.resources.ZLResource;
-import org.geometerplus.zlibrary.ui.android.library.ZLAndroidApplication;
-import org.geometerplus.zlibrary.ui.android.library.ZLAndroidLibrary;
 
 abstract class ZLPreferenceActivity extends SettingsPreferencesActivity {
 	public static String SCREEN_KEY = "screen";
 
-	private Toolbar toolbar;
+
 	private final HashMap<String,Screen> myScreenMap = new HashMap<String,Screen>();
 
 	protected class Screen {
@@ -138,14 +125,13 @@ abstract class ZLPreferenceActivity extends SettingsPreferencesActivity {
 
 	protected abstract void init(Intent intent);
 
+	protected abstract void setupActionBar();
+
 	@Override
 	protected void onCreate(Bundle bundle) {
 		super.onCreate(bundle);
 
-		if(this instanceof PreferenceActivity){
-			setContentView(R.layout.settings);
-			setupActionBar();
-		}
+		setupActionBar();
 
 		Thread.setDefaultUncaughtExceptionHandler(new org.geometerplus.zlibrary.ui.android.library.UncaughtExceptionHandler(this));
 
@@ -157,22 +143,6 @@ abstract class ZLPreferenceActivity extends SettingsPreferencesActivity {
 		setPreferenceScreen(screen != null ? screen.myScreen : myScreen);
 	}
 
-	private void setupActionBar() {
-		toolbar = (Toolbar)findViewById(R.id.toolbar);
-		//Toolbar will now take on default Action Bar characteristics
-		toolbar.setTitle(ZLResource.resource("menu").getResource("preferences").getValue());
-
-		setSupportActionBar(toolbar);
-		getSupportActionBar().setHomeButtonEnabled(true);
-		getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-		toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-			@Override
-			public void onClick(View v) {
-				finish();
-			}
-		});
-	}
 
 	@SuppressWarnings("deprecation")
 	@Override
@@ -208,99 +178,4 @@ abstract class ZLPreferenceActivity extends SettingsPreferencesActivity {
 	}
 
 
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		super.onCreateOptionsMenu(menu);
-		getMenuInflater().inflate(R.menu.toolbar_overflow_menu, menu);
-
-		changeLoginState(menu);
-
-		final ZLAndroidApplication application = (ZLAndroidApplication)getApplication();
-		application.myMainWindow.refreshMenu();
-
-		return true;
-	}
-
-	private void changeLoginState(Menu menu) {
-		MenuItem loginMenuItem = menu.findItem(R.id.menu_item_login_bookshare);
-		MenuItem logoutMenuItem = menu.findItem(R.id.menu_item_logout_bookshare);
-
-		final boolean isLoggedintoBookshare = isLoggedintoBookshare();
-		if(isLoggedintoBookshare) {
-			String title = String.format(getString(R.string.signout_button_title_pattern), getCurrentLoggedUsername());
-			logoutMenuItem.setTitle(title);
-		}
-		loginMenuItem.setVisible(!isLoggedintoBookshare);
-		logoutMenuItem.setVisible(isLoggedintoBookshare);
-	}
-
-	protected boolean isLoggedintoBookshare() {
-		SharedPreferences login_preference = PreferenceManager.getDefaultSharedPreferences(this);
-		String username = login_preference.getString(Bookshare_Webservice_Login.USER, "");
-		String password = login_preference.getString(Bookshare_Webservice_Login.PASSWORD, "");
-		if (username == null || username.isEmpty())
-			return false;
-
-		if (password == null || password.isEmpty())
-			return false;
-
-		return true;
-	}
-
-	protected String getCurrentLoggedUsername() {
-		SharedPreferences login_preference = PreferenceManager.getDefaultSharedPreferences(this);
-		String username = login_preference.getString(Bookshare_Webservice_Login.USER, "");
-		return username;
-	}
-
-	@Override
-	public boolean onOptionsItemSelected(MenuItem item) {
-		final ZLAndroidLibrary zlibrary = (ZLAndroidLibrary) ZLibrary.Instance();
-		if (!zlibrary.isKindleFire() && !zlibrary.ShowStatusBarOption.getValue()) {
-			getWindow().clearFlags(WindowManager.LayoutParams.FLAG_FORCE_NOT_FULLSCREEN);
-		}
-
-		String action = findActionForMenuItem(item.getItemId());
-		Object[] params = findParamsForMenuItemAction(item.getItemId());
-		if(params == null){
-			ZLApplication.Instance().doAction(action);
-		}
-		else {
-			ZLApplication.Instance().doAction(action, params);
-		}
-
-		return super.onOptionsItemSelected(item);
-	}
-
-	@NonNull
-	private String findActionForMenuItem(int itemId) {
-		if (itemId == R.id.menu_item_settings)
-			return ActionCode.SHOW_PREFERENCES;
-
-		if (itemId == R.id.menu_item_sync_with_bookshare) {
-			if (isLoggedintoBookshare())
-				return ActionCode.SYNC_WITH_BOOKSHARE;
-
-			return ActionCode.BOOKSHARE;
-		}
-
-		if (itemId == R.id.menu_item_help)
-			return ActionCode.SHOW_HELP;
-
-		if (itemId == R.id.menu_item_about_goread)
-			return ActionCode.ABOUT_GOREAD;
-
-		if (itemId == R.id.menu_item_logout_bookshare)
-			return ActionCode.LOGOUT_BOOKSHARE;
-
-		if (itemId == R.id.menu_item_login_bookshare)
-			return ActionCode.BOOKSHARE;
-		return "";
-	}
-
-	private Object[] findParamsForMenuItemAction(int itemId){
-		if (itemId == R.id.menu_item_sync_with_bookshare)
-			return new Object[]{SyncReadingListsWithBookshareAction.SyncType.USER_ACTIVATED};
-		return null;
-	}
 }
